+++ base.scss
@@ -16,7 +16,7 @@
 @import "./src/stories/dropdown/dropdown";
 @import "./src/stories/footer/footer";
 @import "./src/stories/breadcrumb/breadcrumb";
-@import "./src/stories/story-placeholder/story-placeholder";
+@import "./src/stories/missing-story/missing-story";
 
 @import "./src/styles/scss/shared";
 @import "./src/styles/scss/internal";
