+++ src/stories/missing-story/missing-story.scss
@@ -1,0 +1,3 @@
+.missing-story {
+  color: #d5364a;
+}
