--- src/stories/story-placeholder/story-placeholder.scss
@@ -1,3 +1,0 @@
-.story-placeholder {
-  color: #d5364a;
-}
